prevent privilege esclation by explicitly not allowing the escalate and bind verbs with rbac on the control plane operator
Currently the control-plane-operator can esclate it's privileges as documented in this link: https://kubernetes.io/docs/reference/access-authn-authz/rbac/#privilege-escalation-prevention-and-bootstrapping. This is due to the fact the control plane operator has escalate and bind privileges implicitly included due to the wildcard in the verb section. That means they can go and edit the rbac that gets assigned to it and increase it's scope. That ability should not be allowed and instead any necessary rbac should be explicitly granted to the control-plane-operator through the hypershift operator